requirements.txt
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.5
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
gunicorn==19.9.0
idna==2.8
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
whitenoise==4.1.2
psycopg2==2.7.1